Don't attempt to perform drop when target is the source
When a drag is initiated and then stopped within a single folder,
there's no need to attempt a drop (and trying to do so will simply
result in an error anyway).

Therefore, when a drag is initiated within a listview folder and stopped
within that folder, no drop attempt will occur.

Note that if the source of the drag is an outside application, or
another tab, the drop will still be attempted, even if the source and
destination folders are the same. That matches the behaviour of Windows
Explorer.