Clear drop description when target is source
The drop description normally indicates where the items will be dropped.
However, no description is shown in Windows Explorer when dragging items
within a folder (and the target is that folder).

Therefore, the drop description will be cleared by Explorer++ when items
are dragged within a listview and the target matches the source.