Update TabContainer class to use ShellDropTargetWindow class
This has the advantage that a drop on a tab will be accepted in the same
way as a drop within the tab contents (i.e. within the listview). This
also simplifies the drop code.

Finally, this fixes an issue where closing the last tab wouldn't close
the application (and could actually cause the application to crash).
That was happening because TabContainer inherited from BaseWindow which
has an OnClose() method. That method would be called instead of
Explorerplusplus::OnClose() and could cause the application to crash
(probably due to the call to DefSubclassProc()).