Only block left-click drop within same folder
When dragging using the left mouse button, the drop should be ignored if
it's within the same folder (as no action would take place anyway).
However, when dragging using the right mouse button, a context menu is
shown on drop and that menu shouldn't be blocked.

Therefore, only left-click drops within a folder are blocked now.