Add more checks before calling IContextMenu::GetCommandString()
From crash dumps supplied in issue #316, it appears that
IContextMenu::GetCommandString() is being called with an invalid (and
potentially uninitialized) ID.

While the ID returned from GetMenuItemInfo() may not be valid if the
item is a separator, I believe it should be valid in other cases.
However, the crash dumps indicate that's not the case.

GetMenuItemInfo() is returning success (otherwise
IContextMenu::GetCommandString() would never be called). But the ID
returned doesn't appear to have a valid value (it's well outside the
range passed to IContextMenu::QueryContextMenu()).

This then causes a crash because IContextMenu::GetCommandString()
doesn't appear to validate values it receives.

To try and protected against that, the MENUITEMINFO struct is now
initialized before the call to GetMenuItemInfo() (in case the original
uninitialized ID is what's being returned) and there are range checks on
the ID. That means that an ID won't be passed to
IContextMenu::GetCommandString() unless it's within the range that was
specified via IContextMenu::QueryContextMenu().