Register as a shell window
Registering as a shell window allows the shell to activate and select
the appropriate window (one that's hosting a particular folder) and
request that the window select an item. This then means that the
application will be correctly invoked when calling
SHOpenFolderAndSelectItems().

The way this works is that the main application window is registered as
a shell window, provided the application is set as the default file
manager. When the shell needs to show an item within a folder, the
application window will be found and brought to the foreground (provided
the target folder is open) and the appropriate item will be selected.

Note that, as part of this commit, a dependency on the
Microsoft.Windows.CppWinRT NuGet package has been added. This is done
purely to make implementing COM classes simpler; no Windows runtime APIs
are used.

Specifically, it means that the IUnknown interface doesn't have to be
repeatedly implemented. While AddRef() and Release() can be implemented
in a generic way easily, QueryInterface() is more difficult. Using a
robust and tested implementation is beneficial.

The other alternatives would be:

- Use ATL. It's not used at all in the application currently and uses
  macros extensively.
- Use WRL. C++/WinRT supersedes this, so there's not much point
  considering it.
- Implement a custom solution. As mentioned above, QueryInterface() is
  difficult to implement in a generic way and not doing it in a generic
  way would mean implementing it manually in every class.
- Copy and paste the implementation for the three methods in every
  class.

Ultimately, using C++/WinRT is beneficial, even if the other parts of
the library aren't (currently) used.

This closes #270.