Don't block drop when modifier keys are down
When dragging an item, the drop is ignored if it's in the same folder.
However, that should only happen if the drop would result in a move. By
default, it will. However, modifier keys can change the drop operation
performed (e.g. to a copy or link).

Therefore, the drop will now only be ignored if the ALT/CTRL/SHIFT keys
aren't down when the drop occurs.

This fixes #324.