Fix crash on startup when using openNewTabNextToCurrent setting
When this setting is enabled, new tabs will be opened next to the
currently selected tab. However, when the application is starting, there
won't be any selected tab. This then meant that if the
openNewTabNextToCurrent setting was enabled, an exception would be
thrown, as the attempt to get the currently selected tab would fail.

To fix this, the openNewTabNextToCurrent setting is now ignored if there
is no selected tab.