Fix issue with dropping items when using DropSource class
DropSource::QueryContinueDrag indicates whether a drop should continue,
be cancelled or completed. The method was refactored in
779f3f92fb3eba7dda101551b7b46d0140274dae, however the refactoring wasn't
correct.

When the mouse button used to initiate the drag is released, the drag
should be completed. However, the refactoring only left the two cases in
which the drag would be cancelled or continued. This meant that once you
started a drag, you couldn't complete it.

To fix this, the method now indicates the drag should be completed if
neither the left nor the right mouse buttons are down.

This fixes #341.