Ensure window is on screen at startup
When the application starts, it's possible that the previous window
position may no longer be visible at all. That could be because, for
example, the monitor setup has changed since the application has last
run.

To address this, the default window position is now used if the loaded
position would be entirely off-screen. This strategy is simple, but
potentially surprising to a user, since the size of the window will
likely change and it will move back to the primary monitor. A more
advanced strategy would be to adjust the position and size of the window
so that it's no longer off-screen, but is still close to the position
and size that the user set.

This fixes #355.