Handle file creation/rename notifications, even if the target file doesn't exist
Because file change notifications are received asynchronously, the
target file may not exist by the time the notification is processed.
That can then cause problems when a file is created or renamed.

Previously, when a file creation notification was processed, if the file
didn't exist, it wouldn't be added.

Similarly, when a rename notification was processed, if the target file
didn't exist, the rename wouldn't be performed.

This behaviour meant that file changes weren't always been tracked
properly and could lead to problems like those described in issue #352,
where files were duplicated.

The new behaviour should mean that file updates are tracked properly,
even if a referenced file doesn't exist when one of the updates is being
processed.

This fixes #352.