Handle file renames/updates in a unified way
Previously, there was one function to handle file rename notifications
and a second very similar function to handle file updates. However, when
handling a file rename, other properties of the file can also have
changed since the rename notification was dispatched, meaning that it
makes sense to treat a rename as an update.

Additionally, in the case where a file is updated, then immediately
renamed, it's necessary to update the file's details when processing the
rename notification, since the update notification won't always be
processed (as the file may no longer exist with the original name at the
time when the update notification is processed).

Therefore, file renames and updates are now handled in the same way.