Fix issue with group update performed when processing file update
When a file is updated, the listview is sorted. That can change the
index of the item being updated. The group update uses the item index
and was being performed after sorting, which wouldn't work properly if
the item had been moved to another position.

To fix this, the group update is now performed before sorting the items.