Update item details immediately after rename
Previously, when renaming an item, the updated text would be kept, but
the actual item details wouldn't change until the rename notification
was processed.

This created a period of time where the user would be shown the updated
name, but the item would still refer to the original name internally.
This then meant that attempting to interact with the item during that
time would fail.

To fix this, the item details are now updated immediately once the user
has entered a name.

This fixes #281.