Fix crash when autosizing columns
Previously, when autosizing columns, ListView_SetColumnWidth was called
for the total number of columns, rather than just the number of active
columns. This then caused LVN_GETDISPINFO messages to be sent for
columns that didn't actually exist, which caused a crash.

To fix this, ListView_SetColumnWidth is now only called for each active
column.

This fixes #350.