Update way in which F2/Tab key presses are handled during rename
When an item is being renamed, F2 can be used to switch between filename
and extension, while Tab/Shift + Tab can be used to switch to the
next/previous item.

Since F2 and Tab/Shift + Tab are all registered as accelerators, the way
this was implemented previously was to forward custom key events to the
listview edit control during a rename.

However, doing that has issues. For example, an accelerator will be
disabled if its corresponding menu item is disabled, which could break
the use of that accelerator. Since the rename menu item is only enabled
when the listview has focus, this behaviour meant that cycling between
the filename and extension with F2 didn't work, because the accelerator
was disabled while renaming an item (since technically, the listview
edit control had focus, not the listview).

To fix this, the F2/Tab/Shift + Tab accelerators are now temporarily
removed whilst renaming an item. This allows the key presses to be
handled directly.

Note that the way in which Tab/Shift + Tab key presses are handled has
also changed. Previously, there was custom code to switch to the
next/previous item. However, the listview has built-in functionality to
do that, so that's what's now used instead.

Because that built-in functionality only appears to work when there's at
least a single listview header item (i.e. column) present, that means
that at least one column is now present in every view mode. That
shouldn't have any other effects, since columns are only actually
visible in details mode.

This fixes #293.