Ensure that tab bar is hidden when loading default folder on startup
When the "Always show the tab bar" setting is disabled, the tab bar is
hidden when there's only a single tab. Additionally, when the "Load
default folder" option is selected, only the default folder will be
loaded on startup.

When these options were used together, the tab bar wasn't being hidden
on startup. This is now fixed.

This fixes #314.