Prevent infinite loop when selecting inaccessible folder in treeview
When selecting a folder in the treeview, the current folder isn't
changed immediately. Instead, a timer is used to change the folder a
short time later.

When attempting to navigate into an inaccessible folder, a dialog box
will be shown. That will then cause the message loop to run. Previously,
that would result in the timer being fired again, which would result in
the navigation being attempted again, in an infinite loop.

To fix this, the timer is now killed before the navigation is started.

This issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?f=5&t=3041.