Simplify way drives toolbar stores data
Drives are now stored in a std::set. This is simpler than the previous
method, in which IDs were stored, along with the a mapping from each ID
to an item.

The new method also has the advantage that the std::set is ordered and
represents data exactly as it's shown in the toolbar.