Rework implementation of DrivesToolbar
The DrivesToolbar has now been split up into three different components:

- DriveModel, which maintains a list of sorted drives.
- DrivesToolbarView, which is responsible for interacting with the
  toolbar control.
- DrivesToolbar, which integrates DriveModel with DrivesToolbarView.

This has the advantage of ensuring a much cleaner separation between the
list of drives and their representation in the toolbar control.
Additionally, it should make it much easier to start testing some of
this code (though there are no tests at the moment).