Retrieve bookmark icon whether or not callback is passed in
When retrieving a bookmark icon, a callback can be passed in that will
be invoked when the icon has been retrieved.

Previously, the icon would only be retrieved if the callback was passed
in. Now, the icon is retrieved whether the callback is passed in or not.
This means that in situations where the callback isn't provided (e.g.
when showing the bookmarks menu), the icon will still be retrieved and
cached and can then be used the next time it's needed.