Handle WM_CONTEXTMENU for treeview and listview
Handling this message means that the custom Shift + F10 and VK_APPS
accelerators don't need to be used, which simplifies things a bit.

It also means that the behaviour described in issue #312 (right clicking
an item after after starting a left click) now shows the context menu.

As part of this change, the code that was in place to prevent an item
from being selected if it was right-clicked outside of the label in list
view mode has been removed. This behaviour was inconsistent with how the
other view modes behave (e.g. right-clicking an item in details mode
results in the item being selected, regardless of whether the label is
clicked or not) and can be added back later if necessary.

Also, the handler that was used when right-clicking an icon in the
display window has been removed.

This closes #312.