Revert text in address bar after enter is pressed
This behaviour more closely matches that of Windows Explorer.

The only case in which the text won't be reverted is when transforming
the entered path fails. For example, that currently happens when the
path is longer than MAX_PATH. Leaving the text in the address bar allows
the user to update it and try again.