Fix crash when running as another user
When the application is run as another user, attempting to create an
instance of the CLSID_ShellWindows class fails.

Previously, the CLSID_ShellWindows instance was being created using
winrt::create_instance. That method throws an exception if the instance
can't be created, which meant that the application would crash in that
situation.

To fix that, winrt::try_create_instance is now used instead.
Additionally, it's no longer assumed that the instance was successfully
created.

This issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?f=5&t=2831.