Fix issue with pasting in zip folders
When pasting a file, the IObjectWithSite interface is retrieved.
However, that interface isn't available in zip folders.

Previously, it was assumed that interface would always be available,
which meant that pasting in a zip folder wouldn't work, as the call to
retrieve the interface would fail.

To fix that, it's no longer assumed that the IObjectWithSite interface
will be available.

This fixes #329.