Suppress exceptions generated by boost::log
Creating or writing to the log file may fail. If that happens, boost
will throw an exception. Previously, that meant that the application
would crash if logging was enabled and the log file couldn't be created
or written to.

Now, exceptions generated by boost::log are suppressed, so issues that
arise from attempting to write to the log file will be ignored.