Set longPathAware field in manifest
This field indicates that the application is opting-in to long path
behaviour. Note that the LongPathsEnabled registry key also needs to be
set (by the user) for this field to have any effect.

One practical difference from this change is that it's possible to
create new folders in a location where the path is longer than MAX_PATH
or the path of the new item would be greater than MAX_PATH.

It also means that the path returned by GetCurrentDirectory() is no
longer restricted to MAX_PATH.

However, aside from those things, it appears that setting this field
probably won't have much of an effect otherwise. That's because the
application is heavily reliant on various shell interfaces, which
seemingly aren't affected by the LongPathsEnabled key.

For example, it's not possible to rename an item that has a long path,
since IShellFolder::SetNameOf() fails, regardless of the
LongPathsEnabled key and longPathAware field.

Also, the "New" submenu shown in the background context menu isn't
properly populated for long path folders.

Additionally, the application still has MAX_PATH hardcoded in quite a
few places, so even if some API methods have had the MAX_PATH
restriction removed, it doesn't mean that a longer path can
automatically be used.

It likely still makes sense to set the longPathAware field though, since
it does make some practical difference and MAX_PATH usages can be
removed over time.