Register main window as drop target
The main window doesn't actually accept any drop items, but this has the
advantage that a drag item is show consistently when a drag is taking
place.

Previously, if a drag item had an image, the image wouldn't always be
shown. If the item was dragged over a window that wasn't registered as a
drop target, the image would revert to just a cursor. This meant that,
as you dragged an item over different child windows:

- The image would be shown and the cursor would indicate that a drop was
  possible.
- The image would be shown and the cursor would indicate that a drop
  wasn't possible.
- The image wouldn't be shown and the cursor would indicate that a drop
  wasn't possible.

By registering the main window as a drop target (and calling the
appropriate IDropTargetHelper methods), the image is always shown,
meaning that there's more consistency in the case where you can't
actually perform a drop.