Fix issue with property sheet page dark mode background colors on Windows 11
Previously, WM_CTLCOLORDLG was being used to set the background color
for each of the property sheet pages in the options dialog in dark mode.
However, it appears that message doesn't work correctly in Windows 11
(the original background color would always be used instead).

To fix that, WM_ERASEBKGND is now used instead.

This is part of the work for issue #115.