Fix formatting of pointer in crash handler
When the application crashes, a separate process will be launched to
create a minidump. That process will be passed the address of the
EXCEPTION_POINTERS struct.

Previously, boost::wformat() was being used to format the command line
for the new process. However, as that method takes locale information
into account, it could format the pointer as a number with thousands
separators. That would then cause the new process to exit, as it
wouldn't be able to parse the command line arguments.

To fix this, std::format() is now used instead. That has a dedicated
formatter for pointers, which should produce the expected output.