Call PostQuitMessage() in WM_NCDESTROY
Previously, this function was being called in WM_DESTROY. That meant
that the Explorerplusplus instance wasn't being destroyed and the
background threads weren't being stopped.

That then made it possible for there to be crashes during application
shutdown, if a background thread (e.g. the directory monitoring thread)
tried to invoke a callback.

To fix these sorts of issues, PostQuitMessage() is now called in
WM_NCDESTROY instead. That means that the Explorerplusplus instance is
destroyed and background threads should be stopped before the
application exits.