Check for visibility when restoring dialog positions
Dialog positions are saved and restored between sessions. However, as
the monitor setup can change between sessions, the previous positions
may no longer be valid. In that case, a dialog that's opened may appear
off screen.

To fix that, the position of a dialog will now only be restored if the
dialog appears on screen.

This issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?f=5&t=3084.