Correctly restore window for existing Explorer++ instance
When the "Allow multiple instances" option is unchecked, opening another
Explorer++ instance will result in a tab being opened in the current
instance. The window for the current instance will also be brought to
the foreground.

Previously, the window was always being restored by passing SW_RESTORE
to the ShowWindow() function. However, that then meant that if the
window was maximized, it would be restored to it's normal
(non-maximized) size.

To fix that, the window is only restored when it's minimized, which is
the only situation in which it's necessary.

This fixes #370.