Correctly handle new tab jump list task when application isn't running
There's a task added to the jump list that allows a new tab to be
created. If that task is executed while the application is running, a
new tab will be created in the existing application instance.

As applications can be pinned to the taskbar, it's also possible for the
task to be executed while the application isn't running. Previously,
that wasn't taken into account and executing the task while the
application wasn't running would have no effect.

To fix that, the application will now start normally if the task is
selected and no current instance exists. A new tab will then be added.