Prevent crash on application shutdown in Windows PE
In environments like Windows PE, there is no shell or taskbar. That then
means that the TaskbarButtonCreated window message will never be
received.

Previously, the TaskbarThumbnails class would create the tab proxies and
destroy them, even if the TaskbarButtonCreated message wasn't received.
When the application was shutting down, that would then cause a crash,
as the DestroyTabProxy() method would attempt to use m_pTaskbarList,
which wouldn't have been set up.

To fix that, the TaskbarThumbnails class has been refactored so that the
tab proxies will only be created once the TaskbarButtonCreated message
has been received and the tab/application observers will only be added
at that point as well.

This issue was reported in #369.