Fix issue with empty treeview in Windows PE
Previously, the treeview was using SHGetFolderLocation() with
CSIDL_DESKTOP to retrieve the pidl of the root desktop folder. However,
that method fails on Windows PE, where the filesystem desktop folder
doesn't exist by default. That then meant that the treeview would be
empty in Windows PE.

To fix that, an alternative method is now used to retrieve the pidl of
the root desktop folder.

This issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?f=2&t=1232.