Disable asynchronous file transfer on Windows PE
When running on Windows PE, attempting to paste a file when using
asynchronous mode fails. Having investigated the issue, it's not clear
why that's the case (perhaps a component that's missing from Windows
PE).

Ideally, it would be best to have a fix that's targeted to the precise
cause of the issue, but since the cause isn't known here, asynchronous
transfer is simply disabled when running on Windows PE. That then allows
copy/paste operations to work as expected.