Update method used to show listview background text
When the listview is empty, a message will be shown on the listview
background. Previously, this was implemented by showing a static image
on the listview. However, that has the following downsides:

- The text is static and therefore isn't localised.
- The font and text size are also static.
- The background and foreground colors are static, so don't integrate
  well with dark mode.

To fix some of these issues, the text is now displayed using
LVN_GETEMPTYMARKUP. This method isn't ideal, since there isn't much
control over the text placement and there's no control over the font or
color used, but it does mean the text can be localised.

There does appear to be a bug with how the listview first draws the
text. When opening the application in an empty folder, the text will
appear slightly off-centre (in Windows 11, at least). Performing an
operation that will cause the listview to redraw (e.g. by moving the
selection rectangle over the area) will cause the text to "jump" into
the correct position.

Several alternative options were also considered:

- Dynamically generating an image for use with ListView_SetBkImage().
  The fact the previous implementation used a static image with
  ListView_SetBkImage() caused several problems. Dynamically generating
  the image isn't too difficult and fixes the problems listed above.
  However, when setting a background image using ListView_SetBkImage(),
  along with a dynamically generated HBITMAP, the image moves when the
  selection rectangle moves over it. That appears to be a bug, since (1)
  loading an image via a resource URL (res://) works as expected (the
  image doesn't move at all) and (2) the image doesn't move when double
  buffering is turned off.
  The image doesn't move when shown as a watermark (with the
  LVBKIF_TYPE_WATERMARK flag set), but is then fixed to the bottom right
  of the listview.
  Any alpha channel in the image is also ignored, unless the image is
  set as a watermark.
- Using WM_ERASEBKGND. Drawing text in this way doesn't work with double
  buffering.
- Drawing the text on top of the listview in WM_PAINT. This does work,
  but there are issues with flickering, since the text is always drawn
  on top of the control.
  Additionally, there are issues with the text background. Using a solid
  background behind the text works, but isn't always correct. For
  example, when the selection rectangle moves over the text, the color
  behind the text changes.
  Using a transparent background works around that problem, but there
  are then issues with overdrawing in certain cases. That is, there are
  times when the text can be drawn over the previous text (and those
  situations are hard to avoid), causing the text to appear incorrectly.

Note that when a filter is applied, a message was also being shown on
the listview background. However, that doesn't really work very well,
since the message can be covered by the items in the listview (that is,
showing a background message really only works well when the listview is
empty).

Therefore, the filter applied message is now shown in the status bar
instead. That should be more readable and will also allow the text to be
localized.