Fix crash that may occur during application shutdown
When the application is in the process of shutting down, it's possible
for windows to be destroyed in the following relative order:

1. The tab control.
2. The main toolbar.

If a WM_SETFOCUS message is received by one of the listview controls in
between steps 1 and 2, MainToolbar::OnFocusChanged() will be invoked.
That will then ultimately result in a call to
TabContainer::GetSelectedTab(), which will cause a crash since the
TabContainer instance will have been destroyed when the tab control was
destroyed (in step 1).

To fix this, focus changed events will no longer be broadcast once
application shutdown has started.