Update how pages in options dialog are selected
Previously, the options dialog used property sheet pages in order to
display each set of options. Property sheets use a tab control to
display the list of pages, however, that can become cumbersome to use as
more pages are added. Having a vertical list of pages works better, as
it's easier to navigate, making it easier to add new pages.

Therefore, the list of options pages is now shown in a treeview. This
also allows nested pages to be added, if necessary.

Doing this also means that the options dialog can be made more fully
compliant with dark mode. Styling a tab control for dark mode isn't
straightforward, so that was one element that still appeared in the
standard colors, even if dark mode was turned on. Since a tab control is
no longer used, that's not an issue anymore.

It also makes it easier to potentially add new features, such as the
ability to search through options pages, since the dialog is fully
controlled by the application.