Remove items from treeview map on deletion
When an item in the navigation treeview is collapsed or removed, each of
its children will be removed from the m_itemInfoMap instance.
Previously, the item itself wouldn't be removed from the map on
deletion. That's now done, to avoid gradually leaking items in the map
as the application is used.