Force dark mode when set
Previously, when enabling dark mode, the application would only appear
as partially dark if dark mode was disabled at the system level. That's
because some of the controls would have custom colors set in dark mode
and would appear dark, but other controls (e.g. menus) would be drawn by
the system and appear light.

To fix that, the ForceDark flag is now used with SetPreferredAppMode()
when dark mode is enabled. That then means that dark mode works
consistently, regardless of the system level setting.

Note that SetPreferredAppMode() was only added in Windows 10 1903, so
enabling dark mode in the application in Windows 10 1809 still requires
dark mode to be enabled at the system level.

This is part of the work for issue #115.