Add a "System default" theme option
Selecting this option means that the application will follow the dark
mode setting in Windows. If the default app mode is set to light, the
application will use a light them. If the default app mode is set to
dark, the application will use a dark theme.

This is part of the work for issue #115.