Rework implementation of color rules functionality
As with the changes made to the application toolbar (in
9f8db6b22107d9ad9d4635ded7fd6b5fa60cf52a) and drives toolbar (in
0a86104b321b77ffb58cfee54d001d7466d51d40), this is a fairly large
change.

As part of this change, the set of color rules is now stored in the
ColorRuleModel class. The advantage of doing that is that the model
class is unit tested. Additionally, unit tests for some of the other
color rules functionality have also been added.

This change also better separates individual parts of the functionality.
For example, the CustomizeColorsDialog was previously updating its
listview directly when changing a color rule. Whereas the listview now
updates itself, using a well defined event. That means the listview will
update, no matter how the set of color rules actually changes.