Fix crash that could occur when handling WM_DEVICECHANGE during application shutdown
When the WM_DEVICECHANGE message is received, each tab as well as the
treeview is notified. Previously, if the message was received at a
specific point during application shutdown (after the tab control had
been destroyed, but before the application exited), a crash would occur,
as the WM_DEVICECHANGE would attempt to use the now-destroyed
TabContainer instance.

To fix that, a signal is now emitted instead, and the treeview and tabs
subscribe to that signal. That way, when the application shuts down,
each open tab will disconnect from the signal before being destroyed.

This fixes #388.