Monitor for drive additions/removals in ShellBrowser
When mapping a network drive, no item creation notification will be
sent. Similarly, when disconnecting the drive, no removal notification
will be sent. That's problematic, since it means that if the user maps a
network drive or disconnects it, no updates will occur.

To fix that, the SHCNE_DRIVEADD and SHCNE_DRIVEREMOVED notifications are
now handled. It should be safe to handle these notifications for any
folder, since an item will only be added/removed internally if it's a
child of the current folder.

Note that when a USB drive is plugged in or removed, a SHCNE_UPDATEDIR
notification is sent. Therefore, using SHCNE_DRIVEADD/SHCNE_DRIVEREMOVED
isn't necessary, but won't cause any issues.

The SHCNE_UPDATEDIR notification is also sent when a virtual hard disk
is attached or removed. However, the notification isn't sent if the disk
is formatted (which will result in the disk being removed). So, handling
SHCNE_DRIVEADD/SHCNE_DRIVEREMOVED is useful there as well.

This is part of the work for #106.