Fix handling of rename item on treeview context menu
In 4cf9ad876fcd1e195824023ca8b3bc5a7cd220d1, the treeview code was
updated to only highlight an item on right-click, rather than select it.
However, that then broke the rename menu item, as it would only place
the selected item into rename mode and the selected item might be
different to the item that was right-clicked.

To fix that, the item will now be renamed based on its pidl instead.