Update way in which node data is stored in ShellTreeView
Previously, item data was stored in an unordered_map. Now, each node
stores a list of its child nodes. This better represents the way the
items are actually displayed and makes it easier to delete items (since
an item can just be deleted without having to manually delete all of its
children).