Update way in which navigation parameters are passed
Navigation parameters are now passed in a single struct. This makes it
easier to keep track of the parameters during a navigation request. This
then means that particular operations can be more easily accomplished.

For example, when navigating to a history entry, the previously selected
items will be re-selected. The method that's called when the navigation
completes can now select those items, since the initial parameters are
available and include information about whether or not the navigation
was a history navigation.

It also makes it easier select the previous item when navigating up, for
the same reason.