Fix tab locking behaviour
Locking a tab prevents it from being closed, but should allow it to be
navigated. Previously, the tab acted like its address was locked as
well, which would then cause all navigations to proceed in a new tab.

This has now been fixed, so that navigations within a locked tab will
proceed in that tab.

This issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?f=4&t=2742.