Allow items to be independently sorted while in groups
Previously, sorting and grouping were effectively exclusive. If items
were grouped, the sort mode would be fixed. If a different sort mode was
set, the items would be ungrouped.

Now, the sort mode can be changed while items are grouped. This allows
items to be grouped in one way, but sorted in another.

This closes #381.