Allow tab control to be scrolled when mouse is over the parent window
When the mouse is towards the right edge of the tab control,
WM_MOUSEWHEEL events will be sent to its parent. Previously,
WM_MOUSEWHEEL was only being handled for the tab control itself. This
meant that scrolling with the mousewheel worked while the mouse was over
a tab, but not otherwise.

To fix that, the WM_MOUSEWHEEL message is now handled for the parent as
well.

Additionally, the scroll direction has been reversed, so that scrolling
up moves left and scrolling down moves right.

This fixes #373.