Mark all classes that inherit from winrt::implements as winrt::non_agile
An agile object is an object which can be freely accessed from any
thread and is the default for classes which inherit from
winrt::implements.

Disabling that behaviour is critically important within the application,
as the classes that exist are effectively never going to be agile; that
is, they expect to be called on only one thread.

Not marking the classes as winrt::non_agile was creating issues. For
example, it meant that when dropping an item from Explorer in a folder,
the FolderView methods could be invoked on a background thread, which
absolutely isn't supported. By marking the class as winrt::non_agile,
the methods will only be invoked on the creating thread (i.e. the main
thread), which is the intended behaviour.

As there doesn't appear to be a way to change the default agile status
in winrt, each class that derives from winrt::implements has been
updated to include winrt::non_agile.